started implementing mappings endpoints, use constants for query parameter mapping and adjust ids of mappings and concepts to int64

+++ internal/api/types_gen.go
@@ -106,7 +106,7 @@ type CodeSystemRoleType string
 // Concept defines model for Concept.
 type Concept struct {
 	Code    *string `json:"code,omitempty"`
-	Id      *int    `json:"id,omitempty"`
+	Id      *uint64 `json:"id,omitempty"`
 	Meaning *string `json:"meaning,omitempty"`
 }
